requirements.txt
requests
paho-mqtt
types-paho-mqtt
python-dotenv-vault
azure-eventhub-checkpointstoreblob-aio
azure-eventhub